refactor(project): move useOverflow to internal/hooks
